Fixes typos on source files and adds CI check
