+++ .github/workflows/main.yml
@@ -36,6 +36,9 @@ jobs:
       - name: Check Bash sources
         run: ./scripts/code-style-bash.sh
 
+      - name: Check typos on code
+        uses: crate-ci/typos@v1.14.10
+
   unit-tests:
     runs-on: ubuntu-20.04
     timeout-minutes: 10
