+++ app/build.gradle.kts
@@ -4,7 +4,7 @@ plugins {
 
 dependencies {
     implementation(projects.platform.jvm.coreRest)
-    implementation(projects.platform.android.corePersistance)
+    implementation(projects.platform.android.corePersistence)
     implementation(projects.platform.android.coreNavigator)
     implementation(projects.platform.android.coreAssets)
     implementation(projects.features.facts)
