+++ features/facts/build.gradle.kts
@@ -6,7 +6,7 @@ dependencies {
     implementation(projects.platform.jvm.coreRest)
     implementation(projects.platform.android.coreAssets)
     implementation(projects.platform.android.coreNavigator)
-    implementation(projects.platform.android.corePersistance)
+    implementation(projects.platform.android.corePersistence)
 
     implementation(libs.kotlinx.coroutines.core)
     implementation(libs.kotlinx.coroutines.android)
@@ -24,7 +24,7 @@ dependencies {
     implementation(libs.androidx.recyclerview)
     implementation(libs.google.material.design)
 
-    testImplementation(projects.platform.android.testingPersistance)
+    testImplementation(projects.platform.android.testingPersistence)
     testImplementation(projects.platform.jvm.testingHelpers)
     testImplementation(projects.platform.jvm.testingRest)
 
@@ -38,7 +38,7 @@ dependencies {
     androidTestImplementation(projects.platform.jvm.testingMockserver)
     androidTestImplementation(projects.platform.android.testingApplication)
     androidTestImplementation(projects.platform.android.testingHelpers)
-    androidTestImplementation(projects.platform.android.testingPersistance)
+    androidTestImplementation(projects.platform.android.testingPersistence)
     androidTestImplementation(projects.platform.android.testingScreenshots)
 
     androidTestImplementation(libs.junit4)
