+++ features/facts/src/test/java/io/dotanuki/features/facts/presentation/FactsViewModelTests.kt
@@ -10,7 +10,7 @@ import io.dotanuki.features.facts.presentation.FactsScreenState.Failed
 import io.dotanuki.features.facts.presentation.FactsScreenState.Idle
 import io.dotanuki.features.facts.presentation.FactsScreenState.Loading
 import io.dotanuki.features.facts.presentation.FactsScreenState.Success
-import io.dotanuki.platform.android.testing.persistance.StorageTestHelper
+import io.dotanuki.platform.android.testing.persistence.StorageTestHelper
 import io.dotanuki.platform.jvm.core.rest.HttpNetworkingError
 import io.dotanuki.platform.jvm.testing.rest.RestScenario
 import io.dotanuki.platform.jvm.testing.rest.RestTestHelper
