+++ features/search/src/main/java/io/dotanuki/features/search/di/SearchContext.kt
@@ -1,7 +1,7 @@
 package io.dotanuki.features.search.di
 
-import io.dotanuki.platform.android.core.persistance.LocalStorage
-import io.dotanuki.platform.android.core.persistance.di.LocalStorageFactory
+import io.dotanuki.platform.android.core.persistence.LocalStorage
+import io.dotanuki.platform.android.core.persistence.di.LocalStorageFactory
 import io.dotanuki.platform.jvm.core.rest.RestClient
 import io.dotanuki.platform.jvm.core.rest.di.ApiUrlFactory
 import io.dotanuki.platform.jvm.core.rest.di.RestClientFactory
