+++ platform/android/core-persistence/.gitignore
