+++ platform/android/core-persistence/src/main/java/io/dotanuki/platform/android/core/persistence/PersistenceContextRegistry.kt
@@ -1,0 +1,14 @@
+package io.dotanuki.platform.android.core.persistence
+
+import android.app.Application
+
+object PersistenceContextRegistry {
+
+    private var app: Application? = null
+
+    fun register(target: Application) {
+        app = target
+    }
+
+    fun targetContext() = app ?: error("No Context available in this registry")
+}
