--- platform/android/testing-persistance/build.gradle.kts
@@ -1,12 +1,0 @@
-
-plugins {
-    id("io.dotanuki.gradle.automodule")
-}
-
-dependencies {
-    implementation(projects.platform.android.corePersistance)
-
-    implementation(libs.kotlinx.coroutines.core)
-    implementation(libs.kotlinx.coroutines.jvm)
-    implementation(libs.androidx.test.monitor)
-}
