+++ platform/android/testing-persistence/proguard-rules.pro
